+++ plugins/event-bus.js
